fix casing of client parameters, move internal sqs mechanism to new file
